refactor: Extracted Field, Table, Connection and Translator out of inputs.snmp
