+++ internal/snmp/snmp_mocks_generate.go
