+++ internal/snmp/testdata/gosmi/foo
