+++ internal/snmp/testdata/gosmi/ifPhysAddress
