+++ internal/snmp/testdata/gosmi/ifPhysAddressImports
