+++ internal/snmp/testdata/gosmi/tableMibImports
