+++ internal/snmp/translator_netsnmp.go
