+++ plugins/inputs/snmp/snmp_test.go
@@ -2,10 +2,9 @@
 package snmp
 
 import (
-	"errors"
 	"fmt"
 	"net"
-	"os/exec"
+	"path/filepath"
 	"sync"
 	"testing"
 	"time"
@@ -65,135 +64,65 @@ func (tsc *testSNMPConnection) Reconnect() error {
 var tsc = &testSNMPConnection{
 	host: "tsc",
 	values: map[string]interface{}{
-		".1.0.0.0.1.1.0":     "foo",
-		".1.0.0.0.1.1.1":     []byte("bar"),
-		".1.0.0.0.1.1.2":     []byte(""),
-		".1.0.0.0.1.102":     "bad",
-		".1.0.0.0.1.2.0":     1,
-		".1.0.0.0.1.2.1":     2,
-		".1.0.0.0.1.2.2":     0,
-		".1.0.0.0.1.3.0":     "0.123",
-		".1.0.0.0.1.3.1":     "0.456",
-		".1.0.0.0.1.3.2":     "0.000",
-		".1.0.0.0.1.3.3":     "9.999",
-		".1.0.0.0.1.5.0":     123456,
-		".1.0.0.1.1":         "baz",
-		".1.0.0.1.2":         234,
-		".1.0.0.1.3":         []byte("byte slice"),
-		".1.0.0.2.1.5.0.9.9": 11,
-		".1.0.0.2.1.5.1.9.9": 22,
-		".1.0.0.0.1.6.0":     ".1.0.0.0.1.7",
-		".1.0.0.3.1.1.10":    "instance",
-		".1.0.0.3.1.1.11":    "instance2",
-		".1.0.0.3.1.1.12":    "instance3",
-		".1.0.0.3.1.2.10":    10,
-		".1.0.0.3.1.2.11":    20,
-		".1.0.0.3.1.2.12":    20,
-		".1.0.0.3.1.3.10":    1,
-		".1.0.0.3.1.3.11":    2,
-		".1.0.0.3.1.3.12":    3,
+		".1.0.0.0.1.1.0":         "foo",
+		".1.0.0.0.1.1.1":         []byte("bar"),
+		".1.0.0.0.1.1.2":         []byte(""),
+		".1.0.0.0.1.102":         "bad",
+		".1.0.0.0.1.2.0":         1,
+		".1.0.0.0.1.2.1":         2,
+		".1.0.0.0.1.2.2":         0,
+		".1.0.0.0.1.3.0":         "0.123",
+		".1.0.0.0.1.3.1":         "0.456",
+		".1.0.0.0.1.3.2":         "0.000",
+		".1.0.0.0.1.3.3":         "9.999",
+		".1.0.0.0.1.5.0":         123456,
+		".1.0.0.0.1.6.0":         ".1.0.0.0.1.7",
+		".1.0.0.1.1":             "baz",
+		".1.0.0.1.2":             234,
+		".1.0.0.1.3":             []byte("byte slice"),
+		".1.0.0.2.1.5.0.9.9":     11,
+		".1.0.0.2.1.5.1.9.9":     22,
+		".1.0.0.3.1.1.10":        "instance",
+		".1.0.0.3.1.1.11":        "instance2",
+		".1.0.0.3.1.1.12":        "instance3",
+		".1.0.0.3.1.2.10":        10,
+		".1.0.0.3.1.2.11":        20,
+		".1.0.0.3.1.2.12":        20,
+		".1.0.0.3.1.3.10":        1,
+		".1.0.0.3.1.3.11":        2,
+		".1.0.0.3.1.3.12":        3,
+		".1.3.6.1.2.1.3.1.1.1.0": "foo",
+		".1.3.6.1.2.1.3.1.1.1.1": []byte("bar"),
+		".1.3.6.1.2.1.3.1.1.1.2": []byte(""),
+		".1.3.6.1.2.1.3.1.1.102": "bad",
+		".1.3.6.1.2.1.3.1.1.2.0": 1,
+		".1.3.6.1.2.1.3.1.1.2.1": 2,
+		".1.3.6.1.2.1.3.1.1.2.2": 0,
+		".1.3.6.1.2.1.3.1.1.3.0": "1.3.6.1.2.1.3.1.1.3",
+		".1.3.6.1.2.1.3.1.1.5.0": 123456,
 	},
-}
-
-func TestFieldInit(t *testing.T) {
-	translations := []struct {
-		inputOid           string
-		inputName          string
-		inputConversion    string
-		expectedOid        string
-		expectedName       string
-		expectedConversion string
-	}{
-		{".1.2.3", "foo", "", ".1.2.3", "foo", ""},
-		{".iso.2.3", "foo", "", ".1.2.3", "foo", ""},
-		{".1.0.0.0.1.1", "", "", ".1.0.0.0.1.1", "server", ""},
-		{".1.0.0.0.1.1.0", "", "", ".1.0.0.0.1.1.0", "server.0", ""},
-		{".999", "", "", ".999", ".999", ""},
-		{"TEST::server", "", "", ".1.0.0.0.1.1", "server", ""},
-		{"TEST::server.0", "", "", ".1.0.0.0.1.1.0", "server.0", ""},
-		{"TEST::server", "foo", "", ".1.0.0.0.1.1", "foo", ""},
-		{"IF-MIB::ifPhysAddress.1", "", "", ".1.3.6.1.2.1.2.2.1.6.1", "ifPhysAddress.1", "hwaddr"},
-		{"IF-MIB::ifPhysAddress.1", "", "none", ".1.3.6.1.2.1.2.2.1.6.1", "ifPhysAddress.1", "none"},
-		{"BRIDGE-MIB::dot1dTpFdbAddress.1", "", "", ".1.3.6.1.2.1.17.4.3.1.1.1", "dot1dTpFdbAddress.1", "hwaddr"},
-		{"TCP-MIB::tcpConnectionLocalAddress.1", "", "", ".1.3.6.1.2.1.6.19.1.2.1", "tcpConnectionLocalAddress.1", "ipaddr"},
-	}
-
-	tr := NewNetsnmpTranslator()
-	for _, txl := range translations {
-		f := Field{Oid: txl.inputOid, Name: txl.inputName, Conversion: txl.inputConversion}
-		err := f.init(tr)
-		require.NoError(t, err, "inputOid=%q inputName=%q", txl.inputOid, txl.inputName)
-		require.Equal(t, txl.expectedOid, f.Oid, "inputOid=%q inputName=%q inputConversion=%q", txl.inputOid, txl.inputName, txl.inputConversion)
-		require.Equal(t, txl.expectedName, f.Name, "inputOid=%q inputName=%q inputConversion=%q", txl.inputOid, txl.inputName, txl.inputConversion)
-	}
-}
-
-func TestTableInit(t *testing.T) {
-	tbl := Table{
-		Oid: ".1.0.0.0",
-		Fields: []Field{
-			{Oid: ".999", Name: "foo"},
-			{Oid: "TEST::description", Name: "description", IsTag: true},
-		},
-	}
-	err := tbl.Init(NewNetsnmpTranslator())
-	require.NoError(t, err)
-
-	require.Equal(t, "testTable", tbl.Name)
-
-	require.Len(t, tbl.Fields, 5)
-	require.Contains(t, tbl.Fields, Field{Oid: ".999", Name: "foo", initialized: true})
-	require.Contains(t, tbl.Fields, Field{Oid: ".1.0.0.0.1.1", Name: "server", IsTag: true, initialized: true})
-	require.Contains(t, tbl.Fields, Field{Oid: ".1.0.0.0.1.2", Name: "connections", initialized: true})
-	require.Contains(t, tbl.Fields, Field{Oid: ".1.0.0.0.1.3", Name: "latency", initialized: true})
-	require.Contains(t, tbl.Fields, Field{Oid: ".1.0.0.0.1.4", Name: "description", IsTag: true, initialized: true})
 }
 
 func TestSnmpInit(t *testing.T) {
 	s := &Snmp{
-		Tables: []Table{
-			{Oid: "TEST::testTable"},
-		},
-		Fields: []Field{
-			{Oid: "TEST::hostname"},
-		},
 		ClientConfig: snmp.ClientConfig{
 			Translator: "netsnmp",
 		},
 	}
 
-	err := s.Init()
-	require.NoError(t, err)
-
-	require.Len(t, s.Tables[0].Fields, 4)
-	require.Contains(t, s.Tables[0].Fields, Field{Oid: ".1.0.0.0.1.1", Name: "server", IsTag: true, initialized: true})
-	require.Contains(t, s.Tables[0].Fields, Field{Oid: ".1.0.0.0.1.2", Name: "connections", initialized: true})
-	require.Contains(t, s.Tables[0].Fields, Field{Oid: ".1.0.0.0.1.3", Name: "latency", initialized: true})
-	require.Contains(t, s.Tables[0].Fields, Field{Oid: ".1.0.0.0.1.4", Name: "description", initialized: true})
-
-	require.Equal(t, Field{
-		Oid:         ".1.0.0.1.1",
-		Name:        "hostname",
-		initialized: true,
-	}, s.Fields[0])
+	require.NoError(t, s.Init())
 }
 
 func TestSnmpInit_noTranslate(t *testing.T) {
-	// override execCommand so it returns exec.ErrNotFound
-	defer func(ec func(string, ...string) *exec.Cmd) { execCommand = ec }(execCommand)
-	execCommand = func(_ string, _ ...string) *exec.Cmd {
-		return exec.Command("snmptranslateExecErrNotFound")
-	}
-
 	s := &Snmp{
-		Fields: []Field{
+		Fields: []snmp.Field{
 			{Oid: ".1.1.1.1", Name: "one", IsTag: true},
 			{Oid: ".1.1.1.2", Name: "two"},
 			{Oid: ".1.1.1.3"},
 		},
-		Tables: []Table{
+		Tables: []snmp.Table{
 			{Name: "testing",
-				Fields: []Field{
+				Fields: []snmp.Field{
 					{Oid: ".1.1.1.4", Name: "four", IsTag: true},
 					{Oid: ".1.1.1.5", Name: "five"},
 					{Oid: ".1.1.1.6"},
@@ -234,8 +163,8 @@ func TestSnmpInit_noTranslate(t *testing.T) {
 
 func TestSnmpInit_noName_noOid(t *testing.T) {
 	s := &Snmp{
-		Tables: []Table{
-			{Fields: []Field{
+		Tables: []snmp.Table{
+			{Fields: []snmp.Field{
 				{Oid: ".1.1.1.4", Name: "four", IsTag: true},
 				{Oid: ".1.1.1.5", Name: "five"},
 				{Oid: ".1.1.1.6"},
@@ -243,8 +172,7 @@ func TestSnmpInit_noName_noOid(t *testing.T) {
 		},
 	}
 
-	err := s.Init()
-	require.Error(t, err)
+	require.Error(t, s.Init())
 }
 
 func TestGetSNMPConnection_v2(t *testing.T) {
@@ -258,8 +186,7 @@ func TestGetSNMPConnection_v2(t *testing.T) {
 			Translator: "netsnmp",
 		},
 	}
-	err := s.Init()
-	require.NoError(t, err)
+	require.NoError(t, s.Init())
 
 	gsc, err := s.getConnection(0)
 	require.NoError(t, err)
@@ -600,147 +527,11 @@ func TestGosnmpWrapper_get_retry(t *testing.T) {
 	require.Equal(t, (gs.Retries+1)*2, reqCount)
 }
 
-func TestTableBuild_walk(t *testing.T) {
-	tbl := Table{
-		Name:       "mytable",
-		IndexAsTag: true,
-		Fields: []Field{
-			{
-				Name:  "myfield1",
-				Oid:   ".1.0.0.0.1.1",
-				IsTag: true,
-			},
-			{
-				Name: "myfield2",
-				Oid:  ".1.0.0.0.1.2",
-			},
-			{
-				Name:       "myfield3",
-				Oid:        ".1.0.0.0.1.3",
-				Conversion: "float",
-			},
-			{
-				Name:           "myfield4",
-				Oid:            ".1.0.0.2.1.5",
-				OidIndexSuffix: ".9.9",
-			},
-			{
-				Name:           "myfield5",
-				Oid:            ".1.0.0.2.1.5",
-				OidIndexLength: 1,
-			},
-			{
-				Name:      "myfield6",
-				Oid:       ".1.0.0.0.1.6",
-				Translate: true,
-			},
-			{
-				Name:      "myfield7",
-				Oid:       ".1.0.0.0.1.6",
-				Translate: false,
-			},
-		},
-	}
-
-	tb, err := tbl.Build(tsc, true, NewNetsnmpTranslator())
-	require.NoError(t, err)
-
-	require.Equal(t, "mytable", tb.Name)
-	rtr1 := RTableRow{
-		Tags: map[string]string{
-			"myfield1": "foo",
-			"index":    "0",
-		},
-		Fields: map[string]interface{}{
-			"myfield2": 1,
-			"myfield3": float64(0.123),
-			"myfield4": 11,
-			"myfield5": 11,
-			"myfield6": "testTableEntry.7",
-			"myfield7": ".1.0.0.0.1.7",
-		},
-	}
-	rtr2 := RTableRow{
-		Tags: map[string]string{
-			"myfield1": "bar",
-			"index":    "1",
-		},
-		Fields: map[string]interface{}{
-			"myfield2": 2,
-			"myfield3": float64(0.456),
-			"myfield4": 22,
-			"myfield5": 22,
-		},
-	}
-	rtr3 := RTableRow{
-		Tags: map[string]string{
-			"index": "2",
-		},
-		Fields: map[string]interface{}{
-			"myfield2": 0,
-			"myfield3": float64(0.0),
-		},
-	}
-	rtr4 := RTableRow{
-		Tags: map[string]string{
-			"index": "3",
-		},
-		Fields: map[string]interface{}{
-			"myfield3": float64(9.999),
-		},
-	}
-	require.Len(t, tb.Rows, 4)
-	require.Contains(t, tb.Rows, rtr1)
-	require.Contains(t, tb.Rows, rtr2)
-	require.Contains(t, tb.Rows, rtr3)
-	require.Contains(t, tb.Rows, rtr4)
-}
-
-func TestTableBuild_noWalk(t *testing.T) {
-	tbl := Table{
-		Name: "mytable",
-		Fields: []Field{
-			{
-				Name:  "myfield1",
-				Oid:   ".1.0.0.1.1",
-				IsTag: true,
-			},
-			{
-				Name: "myfield2",
-				Oid:  ".1.0.0.1.2",
-			},
-			{
-				Name:  "myfield3",
-				Oid:   ".1.0.0.1.2",
-				IsTag: true,
-			},
-			{
-				Name: "empty",
-				Oid:  ".1.0.0.0.1.1.2",
-			},
-			{
-				Name: "noexist",
-				Oid:  ".1.2.3.4.5",
-			},
-		},
-	}
-
-	tb, err := tbl.Build(tsc, false, NewNetsnmpTranslator())
-	require.NoError(t, err)
-
-	rtr := RTableRow{
-		Tags:   map[string]string{"myfield1": "baz", "myfield3": "234"},
-		Fields: map[string]interface{}{"myfield2": 234},
-	}
-	require.Len(t, tb.Rows, 1)
-	require.Contains(t, tb.Rows, rtr)
-}
-
 func TestGather(t *testing.T) {
 	s := &Snmp{
 		Agents: []string{"TestGather"},
 		Name:   "mytable",
-		Fields: []Field{
+		Fields: []snmp.Field{
 			{
 				Name:  "myfield1",
 				Oid:   ".1.0.0.1.1",
@@ -755,11 +546,11 @@ func TestGather(t *testing.T) {
 				Oid:  "1.0.0.1.1",
 			},
 		},
-		Tables: []Table{
+		Tables: []snmp.Table{
 			{
 				Name:        "myOtherTable",
 				InheritTags: []string{"myfield1"},
-				Fields: []Field{
+				Fields: []snmp.Field{
 					{
 						Name: "myOtherField",
 						Oid:  ".1.0.0.0.1.5",
@@ -768,7 +559,7 @@ func TestGather(t *testing.T) {
 			},
 		},
 
-		connectionCache: []snmpConnection{
+		connectionCache: []snmp.Connection{
 			tsc,
 		},
 	}
@@ -787,8 +578,7 @@ func TestGather(t *testing.T) {
 	require.Len(t, m.Fields, 2)
 	require.Equal(t, 234, m.Fields["myfield2"])
 	require.Equal(t, "baz", m.Fields["myfield3"])
-	require.False(t, tstart.After(m.Time))
-	require.False(t, tstop.Before(m.Time))
+	require.WithinRange(t, m.Time, tstart, tstop)
 
 	m2 := acc.Metrics[1]
 	require.Equal(t, "myOtherTable", m2.Measurement)
@@ -802,7 +592,7 @@ func TestGather_host(t *testing.T) {
 	s := &Snmp{
 		Agents: []string{"TestGather"},
 		Name:   "mytable",
-		Fields: []Field{
+		Fields: []snmp.Field{
 			{
 				Name:  "host",
 				Oid:   ".1.0.0.1.1",
@@ -814,7 +604,7 @@ func TestGather_host(t *testing.T) {
 			},
 		},
 
-		connectionCache: []snmpConnection{
+		connectionCache: []snmp.Connection{
 			tsc,
 		},
 	}
@@ -828,366 +618,183 @@ func TestGather_host(t *testing.T) {
 	require.Equal(t, "baz", m.Tags["host"])
 }
 
-func TestFieldConvert(t *testing.T) {
-	testTable := []struct {
-		input    interface{}
-		conv     string
-		expected interface{}
-	}{
-		{[]byte("foo"), "", "foo"},
-		{"0.123", "float", float64(0.123)},
-		{[]byte("0.123"), "float", float64(0.123)},
-		{float32(0.123), "float", float64(float32(0.123))},
-		{float64(0.123), "float", float64(0.123)},
-		{float64(0.123123123123), "float", float64(0.123123123123)},
-		{123, "float", float64(123)},
-		{123, "float(0)", float64(123)},
-		{123, "float(4)", float64(0.0123)},
-		{int8(123), "float(3)", float64(0.123)},
-		{int16(123), "float(3)", float64(0.123)},
-		{int32(123), "float(3)", float64(0.123)},
-		{int64(123), "float(3)", float64(0.123)},
-		{uint(123), "float(3)", float64(0.123)},
-		{uint8(123), "float(3)", float64(0.123)},
-		{uint16(123), "float(3)", float64(0.123)},
-		{uint32(123), "float(3)", float64(0.123)},
-		{uint64(123), "float(3)", float64(0.123)},
-		{"123", "int", int64(123)},
-		{[]byte("123"), "int", int64(123)},
-		{"123123123123", "int", int64(123123123123)},
-		{[]byte("123123123123"), "int", int64(123123123123)},
-		{float32(12.3), "int", int64(12)},
-		{float64(12.3), "int", int64(12)},
-		{int(123), "int", int64(123)},
-		{int8(123), "int", int64(123)},
-		{int16(123), "int", int64(123)},
-		{int32(123), "int", int64(123)},
-		{int64(123), "int", int64(123)},
-		{uint(123), "int", int64(123)},
-		{uint8(123), "int", int64(123)},
-		{uint16(123), "int", int64(123)},
-		{uint32(123), "int", int64(123)},
-		{uint64(123), "int", int64(123)},
-		{[]byte("abcdef"), "hwaddr", "61:62:63:64:65:66"},
-		{"abcdef", "hwaddr", "61:62:63:64:65:66"},
-		{[]byte("abcd"), "ipaddr", "97.98.99.100"},
-		{"abcd", "ipaddr", "97.98.99.100"},
-		{[]byte("abcdefghijklmnop"), "ipaddr", "6162:6364:6566:6768:696a:6b6c:6d6e:6f70"},
-		{[]byte{0x00, 0x09, 0x3E, 0xE3, 0xF6, 0xD5, 0x3B, 0x60}, "hextoint:BigEndian:uint64", uint64(2602423610063712)},
-		{[]byte{0x00, 0x09, 0x3E, 0xE3}, "hextoint:BigEndian:uint32", uint32(605923)},
-		{[]byte{0x00, 0x09}, "hextoint:BigEndian:uint16", uint16(9)},
-		{[]byte{0x00, 0x09, 0x3E, 0xE3, 0xF6, 0xD5, 0x3B, 0x60}, "hextoint:LittleEndian:uint64", uint64(6934371307618175232)},
-		{[]byte{0x00, 0x09, 0x3E, 0xE3}, "hextoint:LittleEndian:uint32", uint32(3812493568)},
-		{[]byte{0x00, 0x09}, "hextoint:LittleEndian:uint16", uint16(2304)},
-	}
+func TestSnmpInitGosmi(t *testing.T) {
+	testDataPath, err := filepath.Abs("../../../internal/snmp/testdata/gosmi")
+	require.NoError(t, err)
 
-	for _, tc := range testTable {
-		act, err := fieldConvert(NewNetsnmpTranslator(), tc.conv, gosnmp.SnmpPDU{Value: tc.input})
-		require.NoError(t, err, "input=%T(%v) conv=%s expected=%T(%v)", tc.input, tc.input, tc.conv, tc.expected, tc.expected)
-		require.EqualValues(t, tc.expected, act, "input=%T(%v) conv=%s expected=%T(%v)", tc.input, tc.input, tc.conv, tc.expected, tc.expected)
-	}
-}
-
-func TestSnmpTranslateCache_miss(t *testing.T) {
-	snmpTranslateCaches = nil
-	oid := "IF-MIB::ifPhysAddress.1"
-	mibName, oidNum, oidText, conversion, err := NewNetsnmpTranslator().SnmpTranslate(oid)
-	require.Len(t, snmpTranslateCaches, 1)
-	stc := snmpTranslateCaches[oid]
-	require.NotNil(t, stc)
-	require.Equal(t, mibName, stc.mibName)
-	require.Equal(t, oidNum, stc.oidNum)
-	require.Equal(t, oidText, stc.oidText)
-	require.Equal(t, conversion, stc.conversion)
-	require.Equal(t, err, stc.err)
-}
-
-func TestSnmpTranslateCache_hit(t *testing.T) {
-	snmpTranslateCaches = map[string]snmpTranslateCache{
-		"foo": {
-			mibName:    "a",
-			oidNum:     "b",
-			oidText:    "c",
-			conversion: "d",
-			err:        errors.New("e"),
+	s := &Snmp{
+		Tables: []snmp.Table{
+			{Oid: "RFC1213-MIB::atTable"},
+		},
+		Fields: []snmp.Field{
+			{Oid: "RFC1213-MIB::atPhysAddress"},
+		},
+		ClientConfig: snmp.ClientConfig{
+			Path:       []string{testDataPath},
+			Translator: "gosmi",
 		},
 	}
-	mibName, oidNum, oidText, conversion, err := NewNetsnmpTranslator().SnmpTranslate("foo")
-	require.Equal(t, "a", mibName)
-	require.Equal(t, "b", oidNum)
-	require.Equal(t, "c", oidText)
-	require.Equal(t, "d", conversion)
-	require.Equal(t, errors.New("e"), err)
-	snmpTranslateCaches = nil
+
+	require.NoError(t, s.Init())
+
+	require.Len(t, s.Tables[0].Fields, 3)
+
+	require.Equal(t, ".1.3.6.1.2.1.3.1.1.1", s.Tables[0].Fields[0].Oid)
+	require.Equal(t, "atIfIndex", s.Tables[0].Fields[0].Name)
+	require.True(t, s.Tables[0].Fields[0].IsTag)
+	require.Empty(t, s.Tables[0].Fields[0].Conversion)
+
+	require.Equal(t, ".1.3.6.1.2.1.3.1.1.2", s.Tables[0].Fields[1].Oid)
+	require.Equal(t, "atPhysAddress", s.Tables[0].Fields[1].Name)
+	require.False(t, s.Tables[0].Fields[1].IsTag)
+	require.Equal(t, "hwaddr", s.Tables[0].Fields[1].Conversion)
+
+	require.Equal(t, ".1.3.6.1.2.1.3.1.1.3", s.Tables[0].Fields[2].Oid)
+	require.Equal(t, "atNetAddress", s.Tables[0].Fields[2].Name)
+	require.True(t, s.Tables[0].Fields[2].IsTag)
+	require.Empty(t, s.Tables[0].Fields[2].Conversion)
+
+	require.Equal(t, ".1.3.6.1.2.1.3.1.1.2", s.Fields[0].Oid)
+	require.Equal(t, "atPhysAddress", s.Fields[0].Name)
+	require.False(t, s.Fields[0].IsTag)
+	require.Equal(t, "hwaddr", s.Fields[0].Conversion)
 }
 
-func TestSnmpTableCache_miss(t *testing.T) {
-	snmpTableCaches = nil
-	oid := ".1.0.0.0"
-	mibName, oidNum, oidText, fields, err := NewNetsnmpTranslator().SnmpTable(oid)
-	require.Len(t, snmpTableCaches, 1)
-	stc := snmpTableCaches[oid]
-	require.NotNil(t, stc)
-	require.Equal(t, mibName, stc.mibName)
-	require.Equal(t, oidNum, stc.oidNum)
-	require.Equal(t, oidText, stc.oidText)
-	require.Equal(t, fields, stc.fields)
-	require.Equal(t, err, stc.err)
-}
-
-func TestSnmpTableCache_hit(t *testing.T) {
-	snmpTableCaches = map[string]snmpTableCache{
-		"foo": {
-			mibName: "a",
-			oidNum:  "b",
-			oidText: "c",
-			fields:  []Field{{Name: "d"}},
-			err:     errors.New("e"),
+func TestSnmpInit_noTranslateGosmi(t *testing.T) {
+	s := &Snmp{
+		Fields: []snmp.Field{
+			{Oid: ".9.1.1.1.1", Name: "one", IsTag: true},
+			{Oid: ".9.1.1.1.2", Name: "two"},
+			{Oid: ".9.1.1.1.3"},
+		},
+		Tables: []snmp.Table{
+			{Name: "testing",
+				Fields: []snmp.Field{
+					{Oid: ".9.1.1.1.4", Name: "four", IsTag: true},
+					{Oid: ".9.1.1.1.5", Name: "five"},
+					{Oid: ".9.1.1.1.6"},
+				}},
+		},
+		ClientConfig: snmp.ClientConfig{
+			Path:       []string{},
+			Translator: "gosmi",
 		},
 	}
-	mibName, oidNum, oidText, fields, err := NewNetsnmpTranslator().SnmpTable("foo")
-	require.Equal(t, "a", mibName)
-	require.Equal(t, "b", oidNum)
-	require.Equal(t, "c", oidText)
-	require.Equal(t, []Field{{Name: "d"}}, fields)
-	require.Equal(t, errors.New("e"), err)
+
+	require.NoError(t, s.Init())
+
+	require.Equal(t, ".9.1.1.1.1", s.Fields[0].Oid)
+	require.Equal(t, "one", s.Fields[0].Name)
+	require.True(t, s.Fields[0].IsTag)
+
+	require.Equal(t, ".9.1.1.1.2", s.Fields[1].Oid)
+	require.Equal(t, "two", s.Fields[1].Name)
+	require.False(t, s.Fields[1].IsTag)
+
+	require.Equal(t, ".9.1.1.1.3", s.Fields[2].Oid)
+	require.Equal(t, ".9.1.1.1.3", s.Fields[2].Name)
+	require.False(t, s.Fields[2].IsTag)
+
+	require.Equal(t, ".9.1.1.1.4", s.Tables[0].Fields[0].Oid)
+	require.Equal(t, "four", s.Tables[0].Fields[0].Name)
+	require.True(t, s.Tables[0].Fields[0].IsTag)
+
+	require.Equal(t, ".9.1.1.1.5", s.Tables[0].Fields[1].Oid)
+	require.Equal(t, "five", s.Tables[0].Fields[1].Name)
+	require.False(t, s.Tables[0].Fields[1].IsTag)
+
+	require.Equal(t, ".9.1.1.1.6", s.Tables[0].Fields[2].Oid)
+	require.Equal(t, ".9.1.1.1.6", s.Tables[0].Fields[2].Name)
+	require.False(t, s.Tables[0].Fields[2].IsTag)
 }
 
-func TestTableJoin_walk(t *testing.T) {
-	tbl := Table{
-		Name:       "mytable",
-		IndexAsTag: true,
-		Fields: []Field{
+func TestGatherGosmi(t *testing.T) {
+	s := &Snmp{
+		Agents: []string{"TestGather"},
+		Name:   "mytable",
+		Fields: []snmp.Field{
 			{
 				Name:  "myfield1",
-				Oid:   ".1.0.0.3.1.1",
+				Oid:   ".1.0.0.1.1",
 				IsTag: true,
 			},
 			{
 				Name: "myfield2",
-				Oid:  ".1.0.0.3.1.2",
+				Oid:  ".1.0.0.1.2",
 			},
 			{
-				Name:                "myfield3",
-				Oid:                 ".1.0.0.3.1.3",
-				SecondaryIndexTable: true,
-			},
-			{
-				Name:              "myfield4",
-				Oid:               ".1.0.0.0.1.1",
-				SecondaryIndexUse: true,
-				IsTag:             true,
-			},
-			{
-				Name:              "myfield5",
-				Oid:               ".1.0.0.0.1.2",
-				SecondaryIndexUse: true,
+				Name: "myfield3",
+				Oid:  "1.0.0.1.1",
 			},
 		},
-	}
+		Tables: []snmp.Table{
+			{
+				Name:        "myOtherTable",
+				InheritTags: []string{"myfield1"},
+				Fields: []snmp.Field{
+					{
+						Name: "myOtherField",
+						Oid:  ".1.0.0.0.1.5",
+					},
+				},
+			},
+		},
 
-	tb, err := tbl.Build(tsc, true, NewNetsnmpTranslator())
-	require.NoError(t, err)
+		connectionCache: []snmp.Connection{tsc},
 
-	require.Equal(t, "mytable", tb.Name)
-	rtr1 := RTableRow{
-		Tags: map[string]string{
-			"myfield1": "instance",
-			"myfield4": "bar",
-			"index":    "10",
-		},
-		Fields: map[string]interface{}{
-			"myfield2": 10,
-			"myfield3": 1,
-			"myfield5": 2,
+		ClientConfig: snmp.ClientConfig{
+			Translator: "gosmi",
 		},
 	}
-	rtr2 := RTableRow{
-		Tags: map[string]string{
-			"myfield1": "instance2",
-			"index":    "11",
-		},
-		Fields: map[string]interface{}{
-			"myfield2": 20,
-			"myfield3": 2,
-			"myfield5": 0,
-		},
-	}
-	rtr3 := RTableRow{
-		Tags: map[string]string{
-			"myfield1": "instance3",
-			"index":    "12",
-		},
-		Fields: map[string]interface{}{
-			"myfield2": 20,
-			"myfield3": 3,
-		},
-	}
-	require.Len(t, tb.Rows, 3)
-	require.Contains(t, tb.Rows, rtr1)
-	require.Contains(t, tb.Rows, rtr2)
-	require.Contains(t, tb.Rows, rtr3)
+	acc := &testutil.Accumulator{}
+
+	tstart := time.Now()
+	require.NoError(t, s.Gather(acc))
+	tstop := time.Now()
+
+	require.Len(t, acc.Metrics, 2)
+
+	m := acc.Metrics[0]
+	require.Equal(t, "mytable", m.Measurement)
+	require.Equal(t, "tsc", m.Tags[s.AgentHostTag])
+	require.Equal(t, "baz", m.Tags["myfield1"])
+	require.Len(t, m.Fields, 2)
+	require.Equal(t, 234, m.Fields["myfield2"])
+	require.Equal(t, "baz", m.Fields["myfield3"])
+	require.WithinRange(t, m.Time, tstart, tstop)
+
+	m2 := acc.Metrics[1]
+	require.Equal(t, "myOtherTable", m2.Measurement)
+	require.Equal(t, "tsc", m2.Tags[s.AgentHostTag])
+	require.Equal(t, "baz", m2.Tags["myfield1"])
+	require.Len(t, m2.Fields, 1)
+	require.Equal(t, 123456, m2.Fields["myOtherField"])
 }
 
-func TestTableOuterJoin_walk(t *testing.T) {
-	tbl := Table{
-		Name:       "mytable",
-		IndexAsTag: true,
-		Fields: []Field{
+func TestGather_hostGosmi(t *testing.T) {
+	s := &Snmp{
+		Agents: []string{"TestGather"},
+		Name:   "mytable",
+		Fields: []snmp.Field{
 			{
-				Name:  "myfield1",
-				Oid:   ".1.0.0.3.1.1",
+				Name:  "host",
+				Oid:   ".1.0.0.1.1",
 				IsTag: true,
 			},
 			{
 				Name: "myfield2",
-				Oid:  ".1.0.0.3.1.2",
-			},
-			{
-				Name:                "myfield3",
-				Oid:                 ".1.0.0.3.1.3",
-				SecondaryIndexTable: true,
-				SecondaryOuterJoin:  true,
-			},
-			{
-				Name:              "myfield4",
-				Oid:               ".1.0.0.0.1.1",
-				SecondaryIndexUse: true,
-				IsTag:             true,
-			},
-			{
-				Name:              "myfield5",
-				Oid:               ".1.0.0.0.1.2",
-				SecondaryIndexUse: true,
+				Oid:  ".1.0.0.1.2",
 			},
 		},
+
+		connectionCache: []snmp.Connection{tsc},
 	}
 
-	tb, err := tbl.Build(tsc, true, NewNetsnmpTranslator())
-	require.NoError(t, err)
+	acc := &testutil.Accumulator{}
 
-	require.Equal(t, "mytable", tb.Name)
-	rtr1 := RTableRow{
-		Tags: map[string]string{
-			"myfield1": "instance",
-			"myfield4": "bar",
-			"index":    "10",
-		},
-		Fields: map[string]interface{}{
-			"myfield2": 10,
-			"myfield3": 1,
-			"myfield5": 2,
-		},
-	}
-	rtr2 := RTableRow{
-		Tags: map[string]string{
-			"myfield1": "instance2",
-			"index":    "11",
-		},
-		Fields: map[string]interface{}{
-			"myfield2": 20,
-			"myfield3": 2,
-			"myfield5": 0,
-		},
-	}
-	rtr3 := RTableRow{
-		Tags: map[string]string{
-			"myfield1": "instance3",
-			"index":    "12",
-		},
-		Fields: map[string]interface{}{
-			"myfield2": 20,
-			"myfield3": 3,
-		},
-	}
-	rtr4 := RTableRow{
-		Tags: map[string]string{
-			"index":    "Secondary.0",
-			"myfield4": "foo",
-		},
-		Fields: map[string]interface{}{
-			"myfield5": 1,
-		},
-	}
-	require.Len(t, tb.Rows, 4)
-	require.Contains(t, tb.Rows, rtr1)
-	require.Contains(t, tb.Rows, rtr2)
-	require.Contains(t, tb.Rows, rtr3)
-	require.Contains(t, tb.Rows, rtr4)
-}
+	require.NoError(t, s.Gather(acc))
 
-func TestTableJoinNoIndexAsTag_walk(t *testing.T) {
-	tbl := Table{
-		Name:       "mytable",
-		IndexAsTag: false,
-		Fields: []Field{
-			{
-				Name:  "myfield1",
-				Oid:   ".1.0.0.3.1.1",
-				IsTag: true,
-			},
-			{
-				Name: "myfield2",
-				Oid:  ".1.0.0.3.1.2",
-			},
-			{
-				Name:                "myfield3",
-				Oid:                 ".1.0.0.3.1.3",
-				SecondaryIndexTable: true,
-			},
-			{
-				Name:              "myfield4",
-				Oid:               ".1.0.0.0.1.1",
-				SecondaryIndexUse: true,
-				IsTag:             true,
-			},
-			{
-				Name:              "myfield5",
-				Oid:               ".1.0.0.0.1.2",
-				SecondaryIndexUse: true,
-			},
-		},
-	}
-
-	tb, err := tbl.Build(tsc, true, NewNetsnmpTranslator())
-	require.NoError(t, err)
-
-	require.Equal(t, "mytable", tb.Name)
-	rtr1 := RTableRow{
-		Tags: map[string]string{
-			"myfield1": "instance",
-			"myfield4": "bar",
-			//"index":    "10",
-		},
-		Fields: map[string]interface{}{
-			"myfield2": 10,
-			"myfield3": 1,
-			"myfield5": 2,
-		},
-	}
-	rtr2 := RTableRow{
-		Tags: map[string]string{
-			"myfield1": "instance2",
-			//"index":    "11",
-		},
-		Fields: map[string]interface{}{
-			"myfield2": 20,
-			"myfield3": 2,
-			"myfield5": 0,
-		},
-	}
-	rtr3 := RTableRow{
-		Tags: map[string]string{
-			"myfield1": "instance3",
-			//"index":    "12",
-		},
-		Fields: map[string]interface{}{
-			"myfield2": 20,
-			"myfield3": 3,
-		},
-	}
-	require.Len(t, tb.Rows, 3)
-	require.Contains(t, tb.Rows, rtr1)
-	require.Contains(t, tb.Rows, rtr2)
-	require.Contains(t, tb.Rows, rtr3)
+	require.Len(t, acc.Metrics, 1)
+	m := acc.Metrics[0]
+	require.Equal(t, "baz", m.Tags["host"])
 }
